fix: compile errors from OPT7/OPT6/OPT2 (CI run 33881573660)
- ConnectionWarmer: HttpUrl.path() is EncodedPath-only on OkHttp 4 ->
  encodedPath("/"); Request.Builder.head() doesn't exist in this
  API surface -> method("HEAD", null).
- GzipRequestInterceptor: RequestBody.isOneShot/isDuplex are FUNCTIONS
  on the OkHttp 4 abstract class, not properties -> add parens.
- ChatViewModel: AutoRetryScheduler.delaySec returns Long (Retry-After
  may be minutes); the countdown loop now ranges Long and narrows to
  Int only at the StateFlow boundary.

+++ src/android/app/src/main/java/com/openminis/app/ui/chat/ChatViewModel.kt
@@ -7775,8 +7775,11 @@ class ChatViewModel(
                             setTransientInlineError("$errSummary — retrying ($retryAttempt/${AUTO_RETRY_DELAYS_SEC.size})…")
                         }
                         try {
-                            for (remaining in delaySec downTo 1) {
-                                _autoRetryCountdown.value = remaining
+                            // [OPT2-jitter-retry-after] delaySec is Long
+                            // (Retry-After can be minutes); countdown var
+                            // stays Long to match.
+                            for (remaining in delaySec downTo 1L) {
+                                _autoRetryCountdown.value = remaining.toInt()
                                 kotlinx.coroutines.delay(1000)
                             }
                         } finally {
